string-based type discriminators in the API
This commit marks the first step in the migration towards string-based
type discriminators in API requests and responses.

String-based discriminators are implemented through additional
`*_type_name` fields. Currently, this doesn't imply a breaking change to
the HTTP API, as both string- and integer-based discriminators are still
supported. Support for integer-based discriminators will be removed in
the v2 version of the API.

At the library level, this change introduces new errors for
(un)marshaling issues. As such, users are advised to review the sentinel
error definitions.